Fixed the structures (I Hope) TODO: Fix attach files stuff
