+++ src/commands/Bind.ts
@@ -2,7 +2,7 @@ module.exports = {
     name: 'bind',
     desc: 'Binds the bot to the channel this is called in',
     execute(client, message, args) {
-        const server = message.guild
+        const server = client.serverManager.addServer(message.guild.id)
         if (!server.gameExists()) {
             server.currentChannel = message.channel
             server.currentChannel.send(
