+++ tsconfig.json
@@ -10,6 +10,7 @@
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "baseUrl": ".",
+      "resolveJsonModule": true,
     "lib": [
       "es5",
       "es2015.promise",
